+++ src/aceinna/devices/upgrade_workers/ethernet_sdk_9100_worker.py
@@ -1041,12 +1041,12 @@ class SDKUpgradeWorker(UpgradeWorkerBase):
 
         self._communicator.reset_buffer()
 
-        for i in range(3):
+        for i in range(6):
             self.send_packet([], send_method=JG)
-            time.sleep(5)
+            time.sleep(3)
             response = helper.read_untils_have_data(
                 self._communicator, JG, 10, 1000)
-            if response:
+            if response is not None:
                 break
 
         if response is not None:
@@ -1175,7 +1175,7 @@ class SDKUpgradeWorker(UpgradeWorkerBase):
         for i in range(3):
             self.send_packet(write_cmd)
             time.sleep(2)
-            result = self.read_until(0xCC, 200, 1)
+            result = self.read_until(0xCC, 1000, 1)
             if result:
                 break
 
@@ -1338,26 +1338,26 @@ class SDKUpgradeWorker(UpgradeWorkerBase):
 
         # need a device ping command, ping 5 times
         self._communicator.reset_buffer()
-        for i in range(3):
+        for i in range(50):
             self.write_wrapper(
                 bytes([int(x, 16) for x in 'ff:ff:ff:ff:ff:ff'.split(':')]),
                 self._communicator.get_src_mac(), pG, [])
             time.sleep(0.2)
             response = helper.read_untils_have_data(
-                self._communicator, pG, 10, 200)
+                self._communicator, pG, 10, 1000)
             if response:
                 break
 
         if not self.send_sdk_cmd_JS():
             return self._raise_error('Send sdk JS command failed')
 
-        for i in range(3):
+        for i in range(50):
             self.write_wrapper(
                 bytes([int(x, 16) for x in 'ff:ff:ff:ff:ff:ff'.split(':')]),
                 self._communicator.get_src_mac(), pG, [])
             time.sleep(0.2)
             response = helper.read_untils_have_data(
-                self._communicator, pG, 10, 200)
+                self._communicator, pG, 10, 1000)
             if response:
                 break